# apps/backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/login:
    post:
      operationId: AuthController_login
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      tags:
        - Auth
  "/auth/register-light":
    post:
      operationId: AuthController_registerLight
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterLightDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      tags:
        - Auth
  "/auth/set-password":
    post:
      operationId: AuthController_setPassword
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      tags:
        - Auth
...
  /email/confirm:
    post:
      operationId: EmailController_confirmEmail
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmEmailDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      tags:
        - Email
  "/email/resend-confirm":
    post:
      operationId: EmailController_resendConfirmationMail
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponseDto"
      tags:
        - Email
  /email/password:
    post:
      operationId: EmailController_sendPasswordLink
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponseDto"
      tags:
        - Email
  "/email/password-change":
    patch:
      operationId: EmailController_changePassword
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponseDto"
      tags:
        - Email
...
  /meetings:
    post:
      operationId: MeetingsController_createMeeting
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeetingDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingResponseDto"
      tags:
        - Meetings
  /meetings/me:
    get:
      operationId: MeetingsController_getMyMeetings
      parameters:
        -
          name: scope
          required: false
          in: query
          schema:
            $ref: "#/components/schemas/MeetingScopeEnum"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeetingResponseDto"
      tags:
        - Meetings
  /meetings/doc:
    get:
      operationId: MeetingsController_getDocMeetingsList
      parameters:
        -
          name: scope
          required: true
          in: query
          schema:
            type: string
        -
          name: page
          required: true
          in: query
          schema:
            type: string
        -
          name: limit
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocMeetingsListResponseDto"
      tags:
        - Meetings
  "/meetings/{id}/cancel":
    post:
      operationId: MeetingsController_cancelMyMeeting
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingResponseDto"
      tags:
        - Meetings
  "/meetings/doc/{id}/cancel":
    post:
      operationId: MeetingsController_cancelAsDoc
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingResponseDto"
      tags:
        - Meetings
...
  "/matches/search-with-slots":
    post:
      operationId: MatchesController_searchDocsWithSlots
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchMatchesDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchMatchesResultDto"
      tags:
        - Matches
  /availability:
    post:
      operationId: AvailabilityController_create
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailabilityDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlotEntity"
      tags:
        - Availability
...
  "/availability/doc/{docId}":
    get:
      operationId: AvailabilityController_freeSlots
      parameters:
        -
          name: docId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilitySlotEntity"
      tags:
        - Availability
  "/availability/{slotId}":
    delete:
      operationId: AvailabilityController_remove
      parameters:
        -
          name: slotId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlotEntity"
      tags:
        - Availability
info:
  title: Melius API
  description: Melius backend API
  version: "1.0"
  contact:
    {}
...
servers:
  []
components:
  schemas:
    LoginDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
    Role:
      type: string
      enum:
        - user
        - doc
        - admin
    AuthResponseDto:
      type: object
      properties:
        role:
          allOf:
            -
              $ref: "#/components/schemas/Role"
        userId:
          type: number
      required:
        - role
        - userId
    RegisterLightDto:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        consentTerms:
          type: boolean
        consentAdult:
          type: boolean
        consentHealthData:
          type: boolean
      required:
        - email
        - consentTerms
        - consentAdult
        - consentHealthData
    SetPasswordDto:
      type: object
      properties:
        password:
          type: string
          minLength: 8
      required:
        - password
    OkResponseDto:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    Language:
      type: string
      enum:
        - pl
        - en
...
    ConfirmEmailDto:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    EmailDto:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ChangePasswordDto:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
      required:
        - token
        - password
...
    Profession:
      type: string
      enum:
        - psychologist
        - psychotherapist
        - sexologist
...
    CreateMeetingDto:
      type: object
      properties:
        slotId:
          type: number
        clientMessage:
          type: string
          maxLength: 1000
      required:
        - slotId
    MeetingStatus:
      type: string
      enum:
        - pending
        - confirmed
        - cancelled_by_user
        - cancelled_by_doc
        - cancelled_by_system
        - completed
    MeetingResponseDto:
      type: object
      properties:
        id:
          type: number
        startTime:
          format: "date-time"
          type: string
        endTime:
          format: "date-time"
          type: string
        status:
          allOf:
            -
              $ref: "#/components/schemas/MeetingStatus"
        clientMessage:
          type: string
          nullable: true
        createdAt:
          format: "date-time"
          type: string
        updatedAt:
          format: "date-time"
          type: string
      required:
        - id
        - startTime
        - endTime
        - status
        - clientMessage
        - createdAt
        - updatedAt
    MeetingScopeEnum:
      type: string
      enum:
        - upcoming
        - past
        - all
    DocMeetingListItemDto:
      type: object
      properties:
        id:
          type: number
        startTime:
          type: string
          format: "date-time"
        endTime:
          type: string
          format: "date-time"
        status:
          allOf:
            -
              $ref: "#/components/schemas/MeetingStatus"
        paid:
          type: boolean
        clientName:
          type: string
          nullable: true
        clientMessage:
          type: string
          nullable: true
      required:
        - id
        - startTime
        - endTime
        - status
        - paid
        - clientName
        - clientMessage
    DocMeetingsListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocMeetingListItemDto"
        page:
          type: number
        limit:
          type: number
        total:
          type: number
      required:
        - items
        - page
        - limit
        - total
...
    SearchMatchesDto:
      type: object
      properties:
        problemId:
          type: number
        from:
          type: string
        to:
          type: string
      required:
        - problemId
        - from
        - to
    SearchMatchesUserDto:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        language:
          allOf:
            -
              $ref: "#/components/schemas/Language"
      required:
        - id
        - email
        - language
    AvailabilitySlotEntity:
      type: object
      properties:
        id:
          type: number
        docId:
          type: number
        startTime:
          format: "date-time"
          type: string
        endTime:
          format: "date-time"
          type: string
        booked:
          type: boolean
      required:
        - id
        - docId
        - startTime
        - endTime
        - booked
    SearchMatchesResultDto:
      type: object
      properties:
        id:
          type: number
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profession:
          nullable: true
          allOf:
            -
              $ref: "#/components/schemas/Profession"
        rate:
          type: number
          nullable: true
        ratesLot:
          type: number
          nullable: true
        sessionPricePln:
          type: number
          nullable: true
        isApproved:
          type: boolean
        user:
          $ref: "#/components/schemas/SearchMatchesUserDto"
        slots:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilitySlotEntity"
      required:
        - id
        - firstName
        - lastName
        - profession
        - rate
        - ratesLot
        - sessionPricePln
        - isApproved
        - user
        - slots
    CreateAvailabilityDto:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
      required:
        - startTime
        - endTime